Struct oak_session_wasm::WasmClientSession

pub struct WasmClientSession { /* private fields */ }

Implementations

impl WasmClientSession

pub fn create_unattested_noise_nn_session( ) -> Result<WasmClientSession, JsValue>

pub fn is_open(&self) -> bool

Checks whether session is ready to send and receive encrypted messages. Session becomes ready once remote attestation and crypto handshake have been successfully finished.

pub fn write(&mut self, plaintext: &[u8]) -> Result<(), JsValue>

Encrypts plaintext and adds it to the queue of outgoing messages in the state-machine.

This function can be called multiple times in a row, which will result in multiple outgoing protocol messages being created.

pub fn read(&mut self) -> Result<Option<Vec<u8>>, JsValue>

Attempts to find a message containing ciphertext in the queue of incoming messages, that is held in the state-machine. If one is found decrypts the ciphertext and returns the plaintext.

This function can be called multiple times in a row.

pub fn put_incoming_message( &mut self, incoming_message: &[u8] ) -> Result<PutIncomingMessageResult, JsValue>

Puts a message received from the peer into the state-machine.

The message is a byte-encoded protobuf message of the type type.googleapis.com/oak.session.v1. It may contain ciphertext, attestation evidence, or a handshake step.

pub fn get_outgoing_message(&mut self) -> Result<Option<Vec<u8>>, JsValue>

Gets the next message that needs to be sent to the peer from the state-machine.

The message is a byte-encoded protobuf message of the type type.googleapis.com/oak.session.v1. It may contain ciphertext, attestation evidence, or a handshake step.